+++ src/dsp/DelayNode.cpp
@@ -24,9 +24,6 @@ DelayNode::DelayNode() :
     distortion = loadParam (distTag);
     pitchSt    = loadParam (pitchTag);
     diffAmt    = loadParam (diffTag);
-#if DIFFUSE_FREQ
-    diffFreq   = loadParam (diffFreqTag);
-#endif
     modFreq    = loadParam (modFreqTag);
     delayMod   = loadParam (delayModTag);
     panMod     = loadParam (panModTag);
@@ -63,9 +60,6 @@ void DelayNode::cookParameters (bool force)
         *distortion,
         *pitchSt,
         *diffAmt,
-#if DIFFUSE_FREQ
-        *diffFreq,
-#endif
         *modFreq,
         *delayMod
     }, force);