+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 - Added individual node modulation, for delay time and panning.
+- Added pitch-shifting and diffusion for each delay node.
 - Added LV2 compilation for Linux.
 - Made Graph View scrollable.
 - Added A/B toggle for switching plugin state.